fix(web): render selected cloud region explicitly
