+++ web/src/features/auth/components/AuthCloudRegionSwitch.tsx
@@ -1,23 +1,8 @@
 import { env } from "@/src/env.mjs";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/src/components/ui/select";
-import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
-import {
-  Dialog,
-  DialogBody,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/src/components/ui/dialog";
-import { useLangfuseCloudRegion } from "@/src/features/organizations/hooks";
+import { CloudRegionPicker } from "@/src/features/auth/components/CloudRegionPicker";
 import { getAvailableCloudRegionOptions } from "@/src/features/organizations/cloudRegions";
+import { useLangfuseCloudRegion } from "@/src/features/organizations/hooks";
+import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 
 export function CloudRegionSwitch({
   isSignUpPage,
@@ -35,143 +20,26 @@ export function CloudRegionSwitch({
   const currentRegion = regions.find((region) => region.name === cloudRegion);
 
   return (
-    <div className="bg-card mt-8 -mb-10 rounded-lg px-6 py-6 text-sm sm:mx-auto sm:w-full sm:max-w-[480px] sm:rounded-lg sm:px-10">
-      <div className="flex w-full flex-col gap-2">
-        <div>
-          <span className="text-sm leading-none font-bold">
-            Data Region
-            <DataRegionInfo />
-          </span>
-          {isSignUpPage && cloudRegion === "HIPAA" ? (
-            <p className="text-muted-foreground text-xs">
-              Demo project is not available in the HIPAA data region.
-            </p>
-          ) : null}
-        </div>
-        <Select
-          value={currentRegion?.name}
-          onValueChange={(value) => {
-            const region = regions.find((region) => region.name === value);
-            if (!region) return;
-            capture(
-              "sign_in:cloud_region_switch",
-              {
-                region: region.name,
-              },
-              {
-                send_instantly: true,
-              },
-            );
-            if (region.hostname) {
-              window.location.hostname = region.hostname;
-            }
-          }}
-        >
-          <SelectTrigger className="w-full">
-            <SelectValue />
-          </SelectTrigger>
-          <SelectContent>
-            {regions.map((region) => (
-              <SelectItem key={region.name} value={region.name}>
-                <span className="flex items-center gap-2">
-                  <span
-                    className={
-                      // The HIPAA symbol glyph sits lower than flag emojis in the same font.
-                      region.name === "HIPAA"
-                        ? "translate-y-[-3px] text-xl leading-none"
-                        : "-translate-y-px text-xl leading-none"
-                    }
-                  >
-                    {region.flag}
-                  </span>
-                  <span>{region.name}</span>
-                </span>
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-
-        {cloudRegion === "HIPAA" && (
-          <div className="bg-muted/50 text-muted-foreground mt-2 rounded-md p-3 text-xs">
-            <p>
-              The Business Associate Agreement (BAA) is only effective on the
-              Cloud Pro and Teams plans.{" "}
-              <a
-                href="https://langfuse.com/security/hipaa"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-link hover:text-link-hover underline"
-              >
-                Learn more about HIPAA compliance →
-              </a>
-            </p>
-          </div>
-        )}
-      </div>
-    </div>
+    <CloudRegionPicker
+      regions={regions}
+      selectedRegion={currentRegion}
+      onValueChange={(value) => {
+        const region = regions.find((region) => region.name === value);
+        if (!region) return;
+        capture(
+          "sign_in:cloud_region_switch",
+          {
+            region: region.name,
+          },
+          {
+            send_instantly: true,
+          },
+        );
+        if (region.hostname) {
+          window.location.hostname = region.hostname;
+        }
+      }}
+      isSignUpPage={isSignUpPage}
+    />
   );
 }
-
-const DataRegionInfo = () => (
-  <Dialog>
-    <DialogTrigger asChild>
-      <a
-        href="#"
-        className="text-link hover:text-link-hover ml-1 text-xs"
-        title="What is this?"
-        tabIndex={-1}
-      >
-        (what is this?)
-      </a>
-    </DialogTrigger>
-    <DialogContent>
-      <DialogHeader>
-        <DialogTitle>Data Regions</DialogTitle>
-      </DialogHeader>
-      <DialogBody>
-        <DialogDescription className="flex flex-col gap-2">
-          <p>Langfuse Cloud is available in four data regions:</p>
-          <ul className="list-disc pl-5">
-            <li>US: Oregon (AWS us-west-2)</li>
-            <li>EU: Ireland (AWS eu-west-1)</li>
-            <li>JP: Tokyo (AWS ap-northeast-1)</li>
-            <li>
-              HIPAA: Oregon (AWS us-west-2) - HIPAA-compliant region (available
-              with Pro and Teams plans)
-            </li>
-          </ul>
-          <p>
-            Regions are strictly separated, and no data is shared across
-            regions. Choosing a region close to you can help improve speed and
-            comply with local data residency laws and privacy regulations.
-          </p>
-          <p>
-            You can have accounts in multiple regions. Each region requires a
-            separate subscription.
-          </p>
-          <p>
-            Learn more about{" "}
-            <a
-              href="https://langfuse.com/security/data-regions"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-link hover:text-link-hover underline"
-            >
-              data regions
-            </a>{" "}
-            and{" "}
-            <a
-              href="https://langfuse.com/docs/data-security-privacy"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-link hover:text-link-hover underline"
-            >
-              data security & privacy
-            </a>
-            .
-          </p>
-        </DialogDescription>
-      </DialogBody>
-    </DialogContent>
-  </Dialog>
-);
